% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcDownloadMultiDay.R
\name{cpcDownloadMultiDay}
\alias{cpcDownloadMultiDay}
\title{Download a Date Range from the CPC FTP Server}
\usage{
cpcDownloadMultiDay(start_date, end_date, download_folder = getwd(),
  overwrite = F)
}
\arguments{
\item{start_date}{first date to download; must be a date object}

\item{end_date}{last date to download; must be a date object}

\item{download_folder}{the folder to which to download the files. Defaults to \code{getwd()}}

\item{overwrite}{if TRUE, will overwrite any file with the same name in the \code{download_folder}.
If FALSE, will not overwrite any file -- if a file with the same name already exists, will
not download again.}
}
\value{
returns a \code{data.frame} with columns date and success. The success column indicates whether
  the raw file for that date was successfully downloaded.
}
\description{
This function is not intended to be called directly.
It downloads a date range from the CPC FTP server by using cpcDownloadOneDay inside of a try statement.
Any dates that are not successfully downloaded are tried a second time.
}

