% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcDownloadOneDay.R
\name{cpcDownloadOneDay}
\alias{cpcDownloadOneDay}
\title{Download a single CPC data file}
\usage{
cpcDownloadOneDay(date, download_folder = getwd(), overwrite = FALSE)
}
\arguments{
\item{date}{\strong{R} date object}

\item{download_folder}{the directory to which you want to download the raw files. Defaults to current directory.}

\item{overwrite}{if TRUE, will overwrite an existing file in the directory. If FALSE and a file with the same name exists,
will download anything, but return successful.}
}
\value{
returns TRUE if the data was downloaded successfully and FALSE otherwise
}
\description{
Download CPC Data for a Single Day
}
\details{
This function returns CPC precipitation data for a single date.
Since the server can be a bit touchy, will try up to five times to connect to the server and
  download the data.
}
\author{
Gopi Goteti

James Doss-Gollin
}
