% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcGlobal.R
\name{cpcGlobal}
\alias{cpcGlobal}
\title{Global Values for the CPC Data Set}
\usage{
cpcGlobal()
}
\description{
This function is not intended to be called directly.
Returns a list which contains several parameters useful for other functions so that
they can be saved in a single location.
}
\author{
James Doss-Gollin
}

