% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcMeltArray.R
\name{cpcMeltArray}
\alias{cpcMeltArray}
\title{Convert a CPC Array to a tidy data.table}
\usage{
cpcMeltArray(array)
}
\arguments{
\item{array}{the array to be converted to a tidy data.table}
}
\value{
returns a data.table with columns date, lon, lat, and prcp_mm.
  The prcp_mm colum will have NA values, corresponding to areas over ocean.
  These can be easily eliminated with \code{na.omit()}.
}
\description{
This function takes in an array with indices [lon, lat, time] and with dimnames giving the values and converts
the array to a data.table.
}

