% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcQueryDateRange.R
\name{cpcQueryDateRange}
\alias{cpcQueryDateRange}
\title{Access CPC Data for a Short Date Range}
\usage{
cpcQueryDateRange(start_date, end_date, tidy = TRUE)
}
\arguments{
\item{start_date}{the first data to download. Must be a date object -- see \code{lubridate} package for easy creation of date objects
with the \code{lubridate::ymd()} function.}

\item{end_date}{as with \code{start_date}, must be a date object.}

\item{tidy}{if TRUE, the output data will be a tidy data.table with columns lon, lat, and prcp_mm.
if FALSE, the output data will be a 3D array with indices [lon, lat, time]}
}
\value{
returns a list with two entries: download_success and precip_data.
  download_success is a data.frame with columns date and success, which define whether data was successfully downloaded for
  each date in the range \code{seq(start_date, end_date, 1)}.
  precip_data is a 3D array if tidy is FALSE and a data.table if tidy is TRUE.
}
\description{
This function provides access to CPC data for a date range.
This function is intended to be called directly.
For each date in the date range, it downloads the raw files, reads them to an array,
and deletes the raw files.
By default it returns a tidy data.table, but it can also return a 3D array.
}
