% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcReadNCDF.R
\name{cpcReadNCDF}
\alias{cpcReadNCDF}
\title{Read in CPC Data from File}
\usage{
cpcReadNCDF(start_date, end_date, lat_lims, lon_lims,
  download_folder = getwd(), tidy = T, round_lonlat = TRUE)
}
\arguments{
\item{start_date}{the first date of data to extract. Must be a date class object: see \code{lubridate::ymd} for easy generation.}

\item{end_date}{the last date of data to extract.}

\item{lat_lims}{a vector of length 2 specifying the minimum and maximum latitudes to query}

\item{lon_lims}{a vector of length 2 specifying the minimum and maximum longitudes to query}

\item{download_folder}{the folder containing the \code{.nc} files you wish to read. Should be the same as the \code{download_folder} argument you passed to
\code{cpcYearToNCDF} unless you have moved the files.}

\item{tidy}{if TRUE, returns the data as a tidy \code{data.table}. If FALSE, returns a 3D array indexed by [lon, lat, time].}

\item{round_lonlat}{if TRUE, the longitude and latitude limits you enter will be rounded to the nearest valid values. If FALSE, if the values you enter are not
in the data set, this function will throw an error.}
}
\value{
returns either a \code{data.table} with columns date, lon, lat, and prcp_mm or a three-dimensional array indexed [lon, lat, time]. This choice
  is controlled by the \code{tidy} argument.
}
\description{
This function allows you to query your customized NCDF files generated by cpcYearToNCDF in an efficient manner with minimal worrying
about syntax. In order for these queries to work, you need to first download the relevant data using the \code{cpcYearToNCDF} function
for the years you wish to download. This function allows you to read those files, using the same \code{download_folder} parameter.
It can return either tidy data or a 3D array.
}
