% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcReadRawOneDay.R
\name{cpcReadRawOneDay}
\alias{cpcReadRawOneDay}
\title{Read Read Data from the CPC}
\usage{
cpcReadRawOneDay(date, download_folder = getwd())
}
\arguments{
\item{date}{the date corresponding to the file to read}

\item{download_folder}{the folder containing the files}
}
\value{
returns a matrix with indices [lon, lat]. Values are daily rainfall, in mm.
}
\description{
This function is not intended to be called directly.
It reads data from a raw file downloaded directly from the CPC ftp server and
returns a matrix with dimensions [lon, lat].
Since each raw file corresponds to data from a single date, there is no date dimension.
}
\author{
Gopi Goteti

James Doss-Gollin
}

