% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcYearToNCDF.R
\name{cpcYearToNCDF}
\alias{cpcYearToNCDF}
\title{Build a NCDF File from CPC Data}
\usage{
cpcYearToNCDF(year, download_folder = getwd(), empty_raw = TRUE,
  overwrite = F)
}
\arguments{
\item{year}{the year for which to download. Must be an integer.}

\item{download_folder}{the folder in which to store the .nc file. Defaults to \code{getwd()}.}

\item{empty_raw}{if TRUE, will delete the raw data. Defaults to TRUE}

\item{overwrite}{if TRUE, will proceed even if a `.nc` file for the specified year exists. If FALSE, will not run.}
}
\value{
a data.table with columns date and success
}
\description{
This function builds a NCDF (.nc) file from CPC data for a single year.
To do this it downloads raw data from the CPC ftp server, reads it to an array, and writes that array to
a NCDF file.
It will always build the file for all data for that particular year.
}

