\name{Fn_Download_CPC_Data_ManyDays}
\alias{Fn_Download_CPC_Data_ManyDays}
\title{Download single CPC data file}
\usage{
  Fn_Download_CPC_Data_ManyDays(begYr, begMo, begDay,
    endYr, endMo, endDay)
}
\arguments{
  \item{begYr}{year, integer}

  \item{begMo}{mo month, integer}

  \item{begDay}{day, integer}

  \item{endYr}{year, integer}

  \item{endMo}{mo month, integer}

  \item{endDay}{day, integer}
}
\value{
  something
}
\description{
  function to download single CPC data file from their ftp
  site
}
\author{
  Gopi Goteti
}

