\name{Fn_Download_CPC_Data_OneDay}
\alias{Fn_Download_CPC_Data_OneDay}
\title{Download single CPC data file}
\usage{
  Fn_Download_CPC_Data_OneDay(yr, mo, day)
}
\arguments{
  \item{yr}{year, integer}

  \item{mo}{month, integer}

  \item{day}{day, integer}
}
\value{
  something
}
\description{
  function to download single CPC data file from their ftp
  site
}
\author{
  Gopi Goteti
}

