\name{Fn_Flip_Matrix_Rows}
\alias{Fn_Flip_Matrix_Rows}
\title{Flip Matrix NS}
\usage{
  Fn_Flip_Matrix_Rows(mat)
}
\arguments{
  \item{mat}{matrix}
}
\value{
  matrix, flipped upside down
}
\description{
  function to flip a matrix upside down (change CPC
  orientation from S-N to N-S)
}
\author{
  Gopi Goteti
}

