\name{Fn_Get_Region_Data_For_Plot}
\alias{Fn_Get_Region_Data_For_Plot}
\title{Extract Region}
\usage{
  Fn_Get_Region_Data_For_Plot(inMat, maxLat, minLat,
    maxLon, minLon)
}
\arguments{
  \item{inMat}{}

  \item{maxLat}{}

  \item{minLat}{}

  \item{maxLon}{}

  \item{minLon}{}
}
\value{
  a matrix, subset by specified bounding box
}
\description{
  function to take global data and extract a regional
  subset and then convert it a form used by ggplot
}
\author{
  Gopi Goteti
}

