\name{Fn_Read_CPC_RawData}
\alias{Fn_Read_CPC_RawData}
\title{Download single CPC data file}
\usage{
  Fn_Read_CPC_RawData(yr, mo, day)
}
\arguments{
  \item{yr}{year, integer}

  \item{mo}{mo month, integer}

  \item{day}{day, integer}
}
\value{
  something
}
\description{
  function to read downloaded CPC data
}
\author{
  Gopi Goteti
}

