\name{Fn_Rotate_Matrix}
\alias{Fn_Rotate_Matrix}
\title{Flip Matrix EW}
\usage{
  Fn_Rotate_Matrix(mat)
}
\arguments{
  \item{mat}{matrix}
}
\value{
  matrix, flipped upside down
}
\description{
  function to rotate a matrix 90 degress clockwise for
  plotting only used to counteract the "image" function
  default behavior
}
\author{
  Gopi Goteti
}

